v2.0.6
- No fail if remote server not enabled
- Net check pings IP instead of hostname

+++ scripts/setup.sh
@@ -1,4 +1,4 @@
-while ! $(ping google.com -c 1 > /dev/null); do
+while ! $(ping 8.8.8.8 -c 1 > /dev/null); do
 	echo Waiting for internet...; sleep 5
 done
 
@@ -12,12 +12,14 @@ apt-add-repository -y universe
 apt -y install wimtools sshfs ssh gawk
 fi
 
-srv=$(<srv); mnt=$(<mnt); pwd=$(<srv-pwd)
-addr=$(echo $srv | gawk 'match($0, /^([A-Za-z0-9]+)@([A-Za-z0-9.]+)(:([0-9]+))?/, a) {print (a[4]?"-p "a[4]" ":"")a[2]}')
-srv=$(echo $srv | gawk 'match($0, /^([A-Za-z0-9]+)@([A-Za-z0-9.]+)(:([0-9]+))?/, a) {print (a[4]?"-p "a[4]" ":"")a[1]"@"a[2]}')
+if [ -f srv ]; then
+	srv=$(<srv); mnt=$(<mnt); pwd=$(<srv-pwd)
+	addr=$(echo $srv | gawk 'match($0, /^([A-Za-z0-9]+)@([A-Za-z0-9.]+)(:([0-9]+))?/, a) {print (a[4]?"-p "a[4]" ":"")a[2]}')
+	srv=$(echo $srv | gawk 'match($0, /^([A-Za-z0-9]+)@([A-Za-z0-9.]+)(:([0-9]+))?/, a) {print (a[4]?"-p "a[4]" ":"")a[1]"@"a[2]}')
 
-echo ---- Connecting to $addr...
-pkill sshfs -KILL || true
-fusermount -u /mnt/share || true
-sh -c "mkdir -p /mnt/share ~/.ssh; ssh-keyscan $addr > ~/.ssh/known_hosts"
-echo $pwd | sshfs -o allow_other,reconnect,password_stdin,ServerAliveInterval=10,auto_unmount $srv:$mnt /mnt/share
+	echo ---- Connecting to $addr...
+	pkill sshfs -KILL || true
+	fusermount -u /mnt/share || true
+	sh -c "mkdir -p /mnt/share ~/.ssh; ssh-keyscan $addr > ~/.ssh/known_hosts"
+	echo $pwd | sshfs -o allow_other,reconnect,password_stdin,ServerAliveInterval=10,auto_unmount $srv:$mnt /mnt/share
+fi
